requirements_f.txt
pandas==2.2.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
scikit-learn==1.2.2
seaborn==0.12.2
sklearn-pandas==2.2.0
xgboost==2.0.3
joblib==1.4.0